Update filename in upload repo action
Sets bundle file created in the Upload Git Repository action to be named after the repository. With the github env variable, an added benefit is that the bundle name is explicitly selected instead of with a wildcard. Instructions for setting env variables are [here](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-environment-variable).